feat: Enhance bet parsing with region, timestamp, and 'per number' logic
This commit significantly enhances the bet slip parsing capabilities of `BetCalculator.php`.

- Implements region-based parsing for '澳门' and '香港'.
- Improves timestamp splitting to be position-agnostic.
- Adds a new rule to parse "各数" pattern for zodiac bets, where the cost is applied per individual number.
