--- backend/.htaccess
@@ -1,16 +1,0 @@
-<IfModule mod_rewrite.c>
-  RewriteEngine On
-  RewriteBase /backend/
-
-  # If the request is not for a real file or directory,
-  # route it to the index.php script.
-  RewriteCond %{REQUEST_FILENAME} !-f
-  RewriteCond %{REQUEST_FILENAME} !-d
-
-  # Rewrite the URL, taking the requested path as the 'action' parameter.
-  # For example, a request to '/backend/get_bills' will be internally
-  # routed to '/backend/index.php?action=get_bills'.
-  # The [QSA] flag ensures any existing query string parameters are appended.
-  # The [L] flag stops processing further rules.
-  RewriteRule ^(.*)$ index.php?action=$1 [QSA,L]
-</IfModule>
