+++ backend/actions/login.php
@@ -25,9 +25,6 @@ try {
     $user = $stmt->fetch(PDO::FETCH_ASSOC);
 
     if ($user && password_verify($password, $user['password'])) {
-        // Regenerate session ID to prevent session fixation attacks
-        session_regenerate_id(true);
-
         $_SESSION['user_id'] = $user['id'];
         $_SESSION['user_email'] = $user['email'];
