--- backend/debug_log.txt
@@ -1,6 +1,0 @@
-
---- New Request: /backend/get_lottery_results ---
-
---- New Request: /backend/get_lottery_results ---
-
---- New Request: /backend/get_lottery_results ---
