--- frontend/.env.development
@@ -1,8 +1,0 @@
-# This file defines environment-specific variables for the development server.
-# Vite will automatically load this file when running in 'development' mode (e.g., `npm run dev`).
-
-# Define the base URL for all API requests made by the frontend client.
-# This will prefix all API calls with '/api', e.g., '/api/get_bills'.
-# We will then configure Vite's dev server to proxy all requests from '/api'
-# to the actual backend server, solving CORS issues during development.
-VITE_API_BASE_URL=/api
