+++ frontend/src/hooks/useLotteryData.jsx
@@ -1,5 +1,4 @@
 import { useState, useEffect, useMemo } from 'react';
-import { getLotteryResults, getGameData } from '../api/client';
 
 export function useLotteryData() {
   const [results, setResults] = useState([]);
@@ -12,14 +11,25 @@ export function useLotteryData() {
       setIsLoading(true);
       setError('');
       try {
-        // Use the new API client for fetching data
-        const [lotteryResults, gameData] = await Promise.all([
-          getLotteryResults(),
-          getGameData()
+        const [resultsResponse, gameDataResponse] = await Promise.all([
+          fetch('/get_lottery_results'),
+          fetch('/get_game_data')
         ]);
 
-        setResults(lotteryResults);
-        setColorMap(gameData.colorMap);
+        const resultsData = await resultsResponse.json();
+        const gameData = await gameDataResponse.json();
+
+        if (resultsData.success) {
+          setResults(resultsData.results);
+        } else {
+          throw new Error(resultsData.error || 'Failed to fetch lottery results.');
+        }
+
+        if (gameData.success) {
+          setColorMap(gameData.colorMap);
+        } else {
+          throw new Error(gameData.error || 'Failed to fetch game data.');
+        }
 
       } catch (err) {
         setError(err.message || 'An error occurred while fetching data.');
