+++ frontend/src/index.css
@@ -59,23 +59,3 @@ a:hover {
   color: var(--primary-color-hover);
   text-decoration: underline;
 }
-
-/* Tag Styles for Bet Slips */
-.time-tag {
-  background-color: #e0e0e0;
-  padding: 2px 6px;
-  border-radius: 4px;
-  font-size: 0.8em;
-  margin-right: 8px;
-  color: #333; /* Ensure text is readable */
-}
-
-.region-tag {
-  background-color: #ffc107; /* A distinct amber color */
-  color: #333;
-  padding: 2px 8px;
-  border-radius: 4px;
-  font-size: 0.8em;
-  font-weight: bold;
-  margin-right: 8px;
-}
